adding cloudinary api

+++ src/modules/daybook/views/EntryView.vue
@@ -7,11 +7,24 @@
         <span class="mx-2 fs-4 fw-light">{{ dayMonthYear.year }}</span>
       </div>
       <div>
-        <button v-if="entry.id" @click="onDeleteEntry" class="btn btn-danger mx-2">
+        <input
+          v-show="false"
+          accept="image/png, image/jpeg, image/jpg"
+          ref="imageSelector"
+          type="file"
+          @change="onSelectedImage"
+        />
+
+        <button
+          v-if="entry.id"
+          @click="onDeleteEntry"
+          class="btn btn-danger mx-2"
+        >
           Borrar
           <font-awesome-icon :icon="['fas', 'fa-trash-alt']" />
         </button>
-        <button class="btn btn-primary">
+
+        <button @click="onSelectImage" class="btn btn-primary">
           Subir Imagen
           <font-awesome-icon :icon="['fas', 'fa-upload']" />
         </button>
@@ -32,9 +45,12 @@
     </div>
 
     <img
-      src="https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/6.png"
+      v-if="entry.picture && !localImage"
+      :src="entry.picture"
       alt="entry-picture"
     />
+
+    <img v-if="localImage" :src="localImage" alt="entry-picture" />
 
     <Fab icon="fa-save" @on:click="onSaveEntry" />
   </template>
@@ -43,7 +59,9 @@
 <script>
 import { defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from 'vuex'
+import Swal from 'sweetalert2'
 import getDayMonthYear from '../helpers/getDayMothYear'
+import uploadImage from '../helpers/uploadImage'
 
 export default {
   components: {
@@ -57,7 +75,9 @@ export default {
   },
   data() {
     return {
-      entry: null
+      entry: null,
+      localImage: null,
+      file: null
     }
   },
   computed: {
@@ -69,6 +89,9 @@ export default {
   methods: {
     ...mapActions('journal', ['updateEntry', 'createEntry', 'deleteEntry']),
     loadEntry() {
+      this.localImage = null
+      this.file = null
+
       let entry
 
       if (this.id === 'new') {
@@ -83,16 +106,60 @@ export default {
       }
     },
     async onSaveEntry() {
+      new Swal({
+        title: 'Wait please',
+        allowOutsideClick: false
+      })
+      Swal.showLoading()
+
+      const picture = await uploadImage(this.file)
+      this.entry.picture = picture
+
       if (this.entry.id) {
         await this.updateEntry(this.entry)
       } else {
+        /* const newId = await this.createEntry({ newEntry: this.entry, uploadImage: this.file }) */
         const newId = await this.createEntry(this.entry)
-        this.$router.push({ name: 'daybook-entry', params: { daybookid: newId } })
+        this.$router.push({
+          name: 'daybook-entry',
+          params: { daybookid: newId }
+        })
       }
+
+      Swal.fire('Saved', 'Entry saved successfully', 'success')
     },
     async onDeleteEntry() {
-      await this.deleteEntry(this.entry.id)
-      this.$router.push({ name: 'no-entry' })
+      const { isConfirmed } = await Swal.fire({
+        title: 'Are you sure?',
+        text: 'Confirm',
+        showDenyButton: true,
+        // eslint-disable-next-line quotes
+        confirmButtonText: "Yes I'm sure"
+      })
+
+      if (isConfirmed) {
+        new Swal({
+          title: 'Wait please',
+          allowOutsideClick: false
+        })
+        Swal.showLoading()
+        await this.deleteEntry(this.entry.id)
+        this.$router.push({ name: 'no-entry' })
+        Swal.fire('Deleted', '', 'success')
+      }
+    },
+    onSelectedImage($event) {
+      const file = $event.target.files[0]
+      if (!file) {
+        this.localImage = null
+        this.file = null
+        return
+      }
+      this.file = file
+      this.localImage = URL.createObjectURL(file)
+    },
+    onSelectImage() {
+      this.$refs.imageSelector.click()
     }
   },
   created() {
